+++ squashfsspec/__init__.py
@@ -2,6 +2,7 @@
 from squashfsspec.squashfsspec import (
     OffsetWrapper,
     SquashFSFileSystem,
+    SquashFSStore,
 )
 
-__all__ = ["SquashFSFileSystem", "OffsetWrapper"]
+__all__ = ["SquashFSFileSystem", "OffsetWrapper", "SquashFSStore"]